requirements.txt
termcolor
Flask-Script
docker-py
mysql-connector-repackaged
psutil
yapsy